+++ fmtc/fmtc_test.go
@@ -32,7 +32,14 @@ func (s *FormatSuite) TestColors(c *C) {
 	c.Assert(Sprint("{b}W{!}"), Equals, "\x1b[0;34;49mW\x1b[0m")
 	c.Assert(Sprint("{m}W{!}"), Equals, "\x1b[0;35;49mW\x1b[0m")
 	c.Assert(Sprint("{c}W{!}"), Equals, "\x1b[0;36;49mW\x1b[0m")
-	c.Assert(Sprint("{s}W{!}"), Equals, "\x1b[0;37;49mW\x1b[0m")
+	c.Assert(Sprint("{s}W{!}"), Equals, "\x1b[0;90;49mW\x1b[0m")
+	c.Assert(Sprint("{r-}W{!}"), Equals, "\x1b[0;91;49mW\x1b[0m")
+	c.Assert(Sprint("{g-}W{!}"), Equals, "\x1b[0;92;49mW\x1b[0m")
+	c.Assert(Sprint("{y-}W{!}"), Equals, "\x1b[0;93;49mW\x1b[0m")
+	c.Assert(Sprint("{b-}W{!}"), Equals, "\x1b[0;94;49mW\x1b[0m")
+	c.Assert(Sprint("{m-}W{!}"), Equals, "\x1b[0;95;49mW\x1b[0m")
+	c.Assert(Sprint("{c-}W{!}"), Equals, "\x1b[0;96;49mW\x1b[0m")
+	c.Assert(Sprint("{s-}W{!}"), Equals, "\x1b[0;90;49mW\x1b[0m")
 }
 
 func (s *FormatSuite) TestBackgrounds(c *C) {
@@ -42,7 +49,8 @@ func (s *FormatSuite) TestBackgrounds(c *C) {
 	c.Assert(Sprint("{B}W{!}"), Equals, "\x1b[0;39;44mW\x1b[0m")
 	c.Assert(Sprint("{M}W{!}"), Equals, "\x1b[0;39;45mW\x1b[0m")
 	c.Assert(Sprint("{C}W{!}"), Equals, "\x1b[0;39;46mW\x1b[0m")
-	c.Assert(Sprint("{S}W{!}"), Equals, "\x1b[0;39;47mW\x1b[0m")
+	c.Assert(Sprint("{S}W{!}"), Equals, "\x1b[0;39;100mW\x1b[0m")
+	c.Assert(Sprint("{W}W{!}"), Equals, "\x1b[0;39;107mW\x1b[0m")
 }
 
 func (s *FormatSuite) TestSpecial(c *C) {