+++ changelog.md
@@ -4,6 +4,8 @@
 
 * `[terminal]` Using forked [go.linenoise](https://github.com/essentialkaos/go-linenoise) package instead original
 * `[terminal]` Added hints support from new version of `go.linenoise`
+* `[fmtc]` Changed colors for `{s}` and `{S}` from light grey to dark grey
+* `[fmtc]` Light colors tag (`-`) support
 
 #### v3.4.2